add StorageLocation to BlockBodyWriter
