+++ crates/storage/storage-api/src/chain.rs
@@ -1,4 +1,4 @@
-use crate::DBProvider;
+use crate::{DBProvider, StorageLocation};
 use alloy_primitives::BlockNumber;
 use reth_chainspec::{ChainSpecProvider, EthereumHardforks};
 use reth_db::{
@@ -22,6 +22,7 @@ pub trait BlockBodyWriter<Provider, Body: BlockBody> {
         &self,
         provider: &Provider,
         bodies: Vec<(BlockNumber, Option<Body>)>,
+        write_to: StorageLocation,
     ) -> ProviderResult<()>;
 
     /// Removes all block bodies above the given block number from the database.
@@ -29,6 +30,7 @@ pub trait BlockBodyWriter<Provider, Body: BlockBody> {
         &self,
         provider: &Provider,
         block: BlockNumber,
+        remove_from: StorageLocation,
     ) -> ProviderResult<()>;
 }
 
@@ -87,6 +89,7 @@ where
         &self,
         provider: &Provider,
         bodies: Vec<(u64, Option<reth_primitives::BlockBody>)>,
+        _write_to: StorageLocation,
     ) -> ProviderResult<()> {
         let mut ommers_cursor = provider.tx_ref().cursor_write::<tables::BlockOmmers>()?;
         let mut withdrawals_cursor =
@@ -116,6 +119,7 @@ where
         &self,
         provider: &Provider,
         block: BlockNumber,
+        _remove_from: StorageLocation,
     ) -> ProviderResult<()> {
         provider.tx_ref().unwind_table_by_num::<tables::BlockWithdrawals>(block)?;
         provider.tx_ref().unwind_table_by_num::<tables::BlockOmmers>(block)?;
